css/css-fonts/parsing/font-variant-ligatures-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-variant-ligatures with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-variant-ligatures-prop">
<meta name="assert" content="font-variant-ligatures supports the full grammar 'normal | none | [ <common-lig-values> || <discretionary-lig-values> || <historical-lig-values> || <contextual-alt-values> ]'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("font-variant-ligatures", "normal");
test_valid_value("font-variant-ligatures", "none");

test_valid_value("font-variant-ligatures", "common-ligatures");
test_valid_value("font-variant-ligatures", "no-common-ligatures");

test_valid_value("font-variant-ligatures", "discretionary-ligatures");
test_valid_value("font-variant-ligatures", "no-discretionary-ligatures");

test_valid_value("font-variant-ligatures", "historical-ligatures");
test_valid_value("font-variant-ligatures", "no-historical-ligatures");

test_valid_value("font-variant-ligatures", "contextual");
test_valid_value("font-variant-ligatures", "no-contextual");

test_valid_value("font-variant-ligatures", "no-common-ligatures no-discretionary-ligatures no-historical-ligatures no-contextual");
</script>
</body>
</html>
